[FEATURE] Move installation to dedicated script
The new install script now queries the last release information directly
from the get.typo3.org API. Therefor only the major version needs to
be defined and not longer the full version. The script recognizes the
doc root so it can simply be changed in the DDEV configuration and
no other changes are needed.