[FEATURE] Add support for local extensions (#11)
* Check for `packages` folder and create symlinks

If the `.ddev/config.yaml` contains a `TYPO3_LOCAL_EXTENSIONS` key under `web_environment`, the script assumes there are local TYPO3 extensions that must be symlinked to `${DDEV_DOCROOT}/typo3conf/ext` (much like it is known from composer-based setups).

If `PACKAGES` is defined and has a value assigned, this value represents the folder name.
If `PACKAGES` is defined, but has no value assigned to it, the script uses `packages` as fallback.
If `PACKAGES` is not defined at all, this section is skipped completely.

Created by Simon Gilli on Slack.

* Add documentation for "Support for local extensions"

* [TASK] Streamline local extension setup

* [TASK] Rename env var

Co-authored-by: Simon Gilli <25326036+gilbertsoft@users.noreply.github.com>